+++ projet-code/back/src/main/java/com/simplon/nooky/project/repository/UserRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.simplon.nooky.project.entities.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 		List<User> findByEmailContaining(String email);
-
+		
 		Optional<User> findByEmail(String email);
 }
